requirements.txt
pandas
mysql-connector-python
sqlalchemy
psycopg2-binary